paginacja

+++ src/main/java/com/mikolaj/e_library/controller/BookController.java
@@ -1,10 +1,15 @@
 package com.mikolaj.e_library.controller;
 
+import com.mikolaj.e_library.DTO.Pagination;
 import com.mikolaj.e_library.DTO.ResponseUtil;
 import com.mikolaj.e_library.DTO.ServiceResponse;
 import com.mikolaj.e_library.model.Book;
 import com.mikolaj.e_library.repo.BookCopyRepository;
 import com.mikolaj.e_library.repo.BookRepository;
+
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,7 +33,39 @@ public class BookController {
         if(response.getData().isEmpty()) return ResponseUtil.okResponse("no activities found", "Activity", Optional.empty());
         return ResponseUtil.okResponse(response.getMessage(), "Books", response.getData());
     }
+    
 
+    @PostMapping("/getAllPaginated")
+    public ResponseEntity<?> getBooks(@RequestBody Pagination pagination){
+        Page<Book> booksPage;
+
+        if (pagination.getFilterBy().isEmpty()) {
+            booksPage = bookRepository.findAll(PageRequest.of(pagination.getPage(), pagination.getSize()));
+        } else if (pagination.getFilterBy().equals("title")) {
+            booksPage = bookRepository.findByTitleContainingIgnoreCase(pagination.getFilter(), PageRequest.of(pagination.getPage(), pagination.getSize(), Sort.by(pagination.getFilterBy()).ascending()));
+        } else {
+            booksPage = bookRepository.findByBookAuthorContainingIgnoreCase(pagination.getFilter(), PageRequest.of(pagination.getPage(), pagination.getSize(), Sort.by(pagination.getFilterBy()).ascending()));
+        }
+
+        if (booksPage.isEmpty()) {
+            return ResponseUtil.okResponse("no books found", "Books", Optional.empty());
+        }
+        return ResponseUtil.okResponse("books found: ", "Books", booksPage);
+    }
+
+    /*
+       {
+        "bookId": 9,
+        "bookType": "kt",
+        "title": "siema",
+        "releaseDate": "2022-01-01",
+        "bookCategory": "bk",
+        "averageBookRating": 5.0,
+        "imageUrl": "costmta.jpg",
+        "bookAuthor": "Rojek",
+        "description": "blebleblbel"
+       }
+     */
     @PostMapping("/save")
     public ResponseEntity<?> saveBook(@RequestBody Book book){
         ServiceResponse<?> response = new ServiceResponse<Object>(Optional.of(bookRepository.save(book)),"book saved");
@@ -36,10 +73,18 @@ public class BookController {
         return ResponseUtil.okResponse(response.getMessage(), "Book", response.getData());
     }
 
+    /*
+       {
+        "bookId": 9,
+       }
+    */
     @DeleteMapping("/delete")
     public ResponseEntity<?> deleteBook(@RequestBody Book book){
-        bookRepository.delete(book);
-        return ResponseUtil.okResponse("book saved", "Book", book);
+        if (bookRepository.findById(book.getBookId()).isPresent()) {
+            bookRepository.delete(book);
+            return ResponseUtil.okResponse("book deleted", "Book", book);
+        }
+        return ResponseUtil.idNotFoundResponse(Book.class);
     }
